Refactor Vault Component: Replace VaultV2Component with VaultComponent and update related imports and routing. Remove obsolete vault-v2 component files.
